requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.0
stripe==7.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.0
gunicorn==21.2.0 